+++ src/grid/GridHeaderComponent.ts
@@ -1,11 +1,12 @@
 import { Output, EventEmitter, ViewChild, TemplateRef, Input } from '@angular/core';
 import {FormioPromiseService} from '../formio-promise.service';
+import {GridHeader} from './types/grid-header';
 
 export class GridHeaderComponent {
   @Input() actionAllowed: any;
-  @Output() sort: EventEmitter<any>;
+  @Output() sort: EventEmitter<GridHeader>;
   @ViewChild(TemplateRef, {static: true}) template: TemplateRef<any>;
-  public headers: Array<any>;
+  public headers: Array<GridHeader>;
   constructor() {
     this.headers = [];
     this.sort = new EventEmitter();