+++ src/grid/form/FormGridHeader.component.ts
@@ -1,17 +1,18 @@
-import { Component } from '@angular/core';
-import { GridHeaderComponent } from '../GridHeaderComponent';
+import {Component} from '@angular/core';
+import {GridHeaderComponent} from '../GridHeaderComponent';
+import {GridHeader, SortType} from '../types/grid-header';
 
 @Component({
   selector: 'form-grid-header',
   templateUrl: './FormGridHeader.component.html'
 })
 export class FormGridHeaderComponent extends GridHeaderComponent {
-  public header: any;
+  public header: GridHeader;
   load(formio?: any) {
     this.header = {
       label: 'Title',
       key: 'title',
-      sort: 'asc'
+      sort: SortType.ASC
     };
     this.headers = [this.header];
     return Promise.resolve(this.headers);