+++ src/grid/types/grid-column.ts
@@ -1,0 +1,7 @@
+import { ComponentInstance } from '../../formio.common';
+
+export interface GridColumn {
+  label?: string;
+  path: string;
+  renderCell?(cellValue: any, component?: ComponentInstance): string;
+}